🩹chore: API 프로토콜 및 KobisOpenAPI 타입 파일 분리
